2018-11-23 Myth
Repentance for expressing the erroneous and false desire
to make one's myth come true. One cannot follow one's own
myth: it only exists in the minds of others. One must be
true to oneself, and if one's contribution is meaningful
and popular, the myth will come from it by itself.